+++ hooks/useMedia/index.mjs
@@ -1,5 +1,5 @@
 /**
- * fetchMedia
+ * useMedia
  *
  * @param {string} mediaId The unique numerical identifier assigned to each image attachment in the media library.
  *
@@ -10,16 +10,16 @@
 import { useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
 
-export function fetchMedia(mediaId) {
+export function useMedia(id) {
 	return useSelect(
 		(select) => {
 			const { getMedia, isResolving, hasFinishedResolution } =
 				select(coreStore);
 
-			const mediaParameters = [mediaId, { context: 'view' }];
+			const mediaParameters = [id, { context: 'view' }];
 
 			return {
-				mediaObj: getMedia(...mediaParameters),
+				media: getMedia(...mediaParameters),
 				isResolvingMedia: isResolving('getMedia', mediaParameters),
 				hasResolvedMedia: hasFinishedResolution(
 					'getMedia',
@@ -27,6 +27,6 @@ export function fetchMedia(mediaId) {
 				),
 			};
 		},
-		[mediaId]
+		[id]
 	);
 }
