+++ index.js
@@ -35,9 +35,9 @@ export { LoadingSpinner } from './components/LoadingSpinner/index.mjs';
 // export { useMedia } from './hooks/useMedia/index.mjs';
 // export { useRenderAppenderWithBlockLimit } from './hooks/useRenderAppenderWithBlockLimit';
 // export { useRequestData } from './hooks/useRequestData';
-export { fetchMedia } from './hooks/fetchMedia/index.mjs';
-export { fetchImage } from './utils/fetchImage/index.mjs';
+export { useMedia } from './hooks/useMedia/index.mjs';
 
 // Utils
 // A utility function is a standard JavaScript function that performs a specific task and is not tied to any particular framework or library. Utility functions are often used for tasks like data formatting, calculations, or other operations that don't require access to React's state or lifecycle. They can be called from anywhere in your code, including within React components or hooks.
 // export { parseMedia } from './utils/parseMedia/index.mjs';
+export { getImageData } from './utils/getImageData/index.mjs';
